# perf: move data out of `Scanner.Token`